Skitter v0.1.0 Skitter.Component.DSL

DSL to define skitter components.

This module offers a DSL to implement components. The entry point of this DSL is the component/3 macro. Besides this, this module offers a set of macros which can be used inside the body of component/3. These macros cannot be used outside component/3, as they rely on AST transformations to work.

A component consists of a collection of metadata alongside the implementation of a set of functions. Metadata is provided in both the header of the component and in its body. Functions are implemented inside the component body. When writing a component, a react/3 function has to be implemented. Besides this, init/3 and terminate/3 can always be implemented (though this is not required), create_checkpoint/3, restore_checkpoint/3 have to be implemented if certain effects are present, clean_checkpoint/3 may be implemented when these effects are present, but this is not required.

A component which stores a running average is provided here as an example:

component Average, in: value, out: average do
  "Track the average of all received values"

  effect state_change
  fields total, counter

  init _ do
    total <~ 0
    counter <~ 0
  end

  react value do
    total <~ total + value
    counter <~ counter + 1

    total / counter ~> average
  end
end

Metadata

The following metadata can be provided when defining a component:

Name	Required?	Location	Syntax

in_ports	always	header	in: [port, another_port]

out_ports	no	header	out: [port, another_port]

description	no	body	"documentation goes here"

effects	no	body	effect effect_name property

Ports define how a component can interact with a workflow. in ports provide a way for the component to receive values. A component instance automatically reacts to data when it has an input value on each of its input ports. out ports allow a component to publish data while reacting to incoming data. Any data published on an out port will automatically be forwarded to any component instance which is connected to this out port. A component must always specify its in ports, out ports do not need to be specified in which case the component cannot publish any data. As a syntactic convenience, a component can leave out the list notation when only one port is present, e.g. in: foo, out: bar is automatically transformed into in: [foo], out: [bar]

A description can be provided as the first line of a component. It simply serves as documentation for the component.

Effects modify how the skitter runtime will handle the execution of a component. Currently, skitter supports the following effects:

Name	Properties

state_change	hidden

external_effect

The state_change effect has to be specified when a component may modify its internal state when reacting to incoming data. Setting this effect will allow the modification of state in react. Furthermore, the skitter runtime will ensure that there are no concurrent modifications to this state when a workflow is executed on a distributed system. Finally, the skitter runtime will also guarantee that the state can be recovered if failure occurs. Skitter may not always be able to “see” the entire state of a component instance; this can be the case if the component relies on an external process to do most of the processing. In this case, a state_change effect can be annotated with the hidden property. When this is done, skitter will rely on custom implementations of checkpoint functionality to ensure the state remains recoverable in the case of partial failure.

The external_effect effect has to be specified when a component may trigger external effects (i.e. I/O) when reacting. When this effect is specified, the after_failure/1 macro may be used inside react/3. This block enables the component developer to ensure that a side effect occurs only once.

Functions

Various functions can be implemented inside the body of component/3. This can be done by using one of the macros defined and documented inside this module.

Fields

Fields can be specified inside the body of component/3 with the following syntax: fields field1, field2. These fields can be used to store state inside a component instance. Values can be assigned to fields inside both init/3 and restore_checkpoint/3 with the following syntax: field1 <~ value. These values can be obtained from the component instance inside every other component function (react/3, terminate/3, create_checkpoint/3 and clean_checkpoint/3) by using the field name. when the state_change effect is present, fields may be modified inside the body of react/3.

Functions

after_failure(list) (macro)

Code that should only be executed after a failure occurred.

Usable inside react/3 iff the component has an external state.

The code in this block will only be executed if react/3 is triggered after a failure occurred. Otherwise, the code in this block will be ignored.

This block is mainly meant to provide clean up code in case a component experiences some form of failure. For instance, if a call to react can produce some side effect, this block can check if that side effect already occurred before deciding whether or not execution should proceed.

Example

For instance, let’s look at the following react function, which writes a value to some database after which it sends it to an output port.

react val do
  write_to_db(val)
  val ~> port
end

The skitter runtime would have no way of knowing whether or not the database was updated if this component would fail in the middle of a call to react. However, in order to ensure the remainder of the workflow is still activated correctly, the value sent to port is needed.

To solve this problem, we can use the after_failure block:

react val do
  after_failure do
    if write_to_db_occurred?(val) do
      val ~> port
      skip
    end
  end

  write_to_db(val)
  val ~> port
end

The after_failure block will only be executed if a previous call to react failed. In this case, it would check if the database was updated. If this is the case, it sends the value to port and it uses skip/2 to ensure the remainder of react is not carried out.

clean_checkpoint(list1, meta, list2) (macro)

Clean up an existing checkpoint.

Clean up a checkpoint that will no longer be used by the skitter runtime. Do not call this macro manually, as it relies on AST transformations to work; use the following syntax instead:

clean_checkpoint arg do
  # Clean up here
end

clean_checkpoint/3 always accepts a single argument, which is the checkpoint that should be removed.

component(name, ports, list) (macro)

Create a skitter component.

This macro serves as the entry point of the component DSL. Don’t call any other macro in this module outside the body of component/3; they rely on AST transformation to work properly.

This macro will generate a skitter component, which is implemented as an elixir module with a set of predefined functions. The name argument to this function will be used as the name of the module. The ports argument is used to define the ports of the component, the body should implement the functionality of the component.

Ports

Ports are provided to the component as a keyword list with two valid keys: in:, and out:, representing the in and out ports respectively. In ports should always be provided while out ports may be omitted. Ports are specified as a list of Elixir names. As a syntactic convenience, the list notation can be omitted if only a single port exists.

The following are all valid port specifications:

in: foo, out: bar

in: foo

in: [foo, bar]

in: foo, out: [bar, baz]

The following port specifications will not be accepted by the component:

out: foo

in: :foo

Body

The following code is allowed inside the body of component/3:

An effect statement, which declares an effect of the component.

A field statement, which specifies the data layout of the state of the component.

A call to the react/3 macro, this is required.

A call to init/3 or terminate/3

A call to create_checkpoint/3 or restore_checkpoint/3, this is required if the component has the hidden state_change effect.

A call to clean_checkpoint/3; this is allowed, but not required when component has the hidden state_change effect.

The definition of a helper function.

Besides this, a description may be provided by providing a string as the first statement in the component body.

Example

component Average, in: value, out: average do
  "Track the average of all received values"

  effect state_change
  fields total, counter

  init _ do
    # implementation goes here...
  end

  react value do
    # implementation goes here...
  end
end

create_checkpoint(list1, meta, list2) (macro)

Create a checkpoint.

This macro should be used to create a checkpoint based on a component instance. Do not call this macro directly as it relies on AST transformations. Use the following syntax instead:

checkpoint do
# body
end

The return value (i.e. the value of the last statement in the body) is returned as the checkpoint to the skitter runtime.

error(reason) (macro)

Stop the current callback and return with an error.

A reason should be provided as a string. In certain contexts (e.g. init/3), the use of this macro will crash the entire workflow. Any spits or state updates that may have occurred prior to the invocation of this macro will be lost.

init(list1, meta, list2) (macro)

Instantiate a skitter component.

Set up the initial state of the component based on a single argument provided by the workflow. Do not call this macro manually, as it relies on AST transformations to work. Instead, use the following syntax:

init arg do
  # Body goes here
end

Init always accepts a single argument, pattern matching can be used to deconstruct complex arguments which may be provided. Inside init, field <~ value may be used to set initial values for the state fields.

react(args, meta, list) (macro)

React to incoming data.

Implement the behaviour of the component when it receives data from the in ports. Do not call this macro manually as it relies on AST transformations to work, instead use the macro with the following syntax:

react arg_1, arg_2, ..., arg_n do
  # Body goes here
  end

The amount of arguments in the arguments list of react should correspond to the amount of in ports of the component. When the component instance needs to react, every argument will be bound to the value provided to its corresponding in port. While reacting, the ~> operator can be used to spit data to any out port; <~ can be used to update the state of the component instance if and only if the component has the state_change effect. after_failure/1 can be used inside react/3 if and only if the component has the external_effect effect.

The return value of react/3 is silently ignored. However, any “spits” or state updates performed during the execution of react/3 will be provided to the skitter runtime.

Spits

While a component instance is reacting, it may spit a value to an out port with the following syntax: value ~> port. The given value will be sent to any in_port that is connected to the out port in question. The values will only be sent after the “current” react is finished. Thus, if error/1 is called after ~>, the value will never be sent to the connected ports.

State Updates

A component with the state_change effect may update any field of the state (defined by using fields inside component/3) with the following syntax: field <~ value. The component instance will immediately be updated to reflect these changes. This implies that later uses of field will contain the new state of the modified field. In spite of this, the skitter runtime will only see these changes once the “current” react is finished.

restore_checkpoint(list1, meta, list2) (macro)

Restore a component instance from a checkpoint.

Similar to init/3; set up the initial state of the component based on a checkpoint argument. Do not call this macro manually, as it relies on AST transformations to work. Instead, use the following syntax:

restore_checkpoint arg do
  # Recreate state here
end

restore_checkpoint always accepts a single argument, which is a checkpoint as returned by create_checkpoint/3. The body of restore_checkpoint/3 is responsible for reconstructing the component state (using the field <~ value syntax) to set correct values for the state fields.

skip(state, output) (macro)

Stop the execution of react, and return the current instance state and spits.

Using this macro will automatically stop the execution of react. Unlike the use of error/1, any state changes and spit values will still be provided to the skitter runtime.

This macro is useful when the execution of react should only continue under certain conditions. It is especially useful in an after_failure/1 block, as it can be used to only continue the execution of react if no effect occurred in the original call to react.

Do not provide arguments when using this macro (i.e. just use skip), the state and output arguments will automatically be injected by react/3.

terminate(list1, meta, list2) (macro)

Generate component clean up code.

This macro can be used to clean up any resources before a component instance is shut down. Do not call this macro manually as it relies on AST transformations to work. Use the following syntax:

terminate do
  # body
end