Skitter v0.1.0 Skitter.Workflow.DSL

DSL to define skitter workflows.

This module offers the workflow/1 macro, which should be used if you plan to write a workflow by hand. If you want to automatically generate a workflow (e.g. based on the input to some graphical tool), you can use this format or the layout described in Skitter.Component. Internally, the macro will compile its input to the same representation. Additionally, this macro will ensure that the workflow does not contain some common errors.

A workflow definition is a list of component instances with a unique name. The following syntax is used: instance_name = {component_module, initialization_arguments, links}. For instance, if we have a component Foo, with in ports a, b, c and out ports: d, e we can define the following workflow:

workflow do
  instance = {
    Foo, _,
    d ~> other.a,
    d ~> other.b,
    e ~> other.c
  }
  other = {Foo, _}
end

This workflow consists of multiple instances of the same component (Foo). Foo does not take any initialisation arguments (which is why _ is used). The first instance of Foo (instance) is linked to the second instance of Foo (other) through the use of the link syntax: out_port ~> instance.in_port. As shown in the example, multiple links can start from the same out port, however, only one incoming link is allowed per input port.

Functions

workflow(list) (macro)

Create a skitter workflow.

This macro serves as the entry point of the Skitter.Workflow.DSL DSL. Please refer to the module documentation for additional details.