Skitter v0.1.0-dev Skitter.Workflow

Documentation for Workflow.

Types

destination()

destination() :: {workflow_identifier(), Skitter.Component.port_name()}

Address a token can be sent to.

Specified as the combination of a valid workflow identifier that refers to a component instance and the name of a valid port of that component.

proto_instance()

proto_instance() ::
  {Skitter.Component.t(), any(),
   [{Skitter.Component.port_name(), [destination()]}]}

Proto-instance type.

A proto-instance is a tuple which contains the component it represents, its init argument (which can be anything), and a list of destinations, arranged by out port.

t()

t() :: %Skitter.Workflow{
  instances: %{required(workflow_identifier()) => proto_instance()},
  sources: %{required(workflow_identifier()) => [destination()]}
}

Workflow data structure.

A workflow is defined as a combination of named sources and proto-instances.

workflow_identifier()

workflow_identifier() :: atom()

Identifier of an instance or source in a workflow.

Functions

get_component(workflow, key)

Examples

iex> get_component(example_workflow(), :i) Identity iex> get_component(example_workflow(), :does_not_exist) ** (KeyError) Key :does_not_exist not found in workflow

get_init(workflow, key)

Examples

iex> get_init(example_workflow(), :i) nil

get_links(workflow, key)

Examples

iex> get_links(example_workflow(), :i) []

workflow(list) (macro)

Create a workflow.