Skitter v0.1.0-dev Skitter.Workflow

Tools to interact with skitter workflows.

This module defines an API that can be used to fetch data from skitter workflow structures. Besides this, this module defines the workflow/1 macro which can be used to define a skitter workflow.

Types

destination()

destination() :: {workflow_identifier(), Skitter.Component.port_name()}

Address a token can be sent to.

Specified as the combination of a valid workflow identifier that refers to a component instance and the name of a valid port of that component.

outgoing_links()

outgoing_links() :: [{Skitter.Component.port_name(), [destination()]}]

Outgoing links of a component

Specified as a keyword list. Each key in the list corresponds to an out port of the component, while the values for this key correspond to all the t:destination ports this out port is connected with.

proto_instance()

proto_instance() :: {Skitter.Component.t(), any(), outgoing_links()}

Proto-instance type.

A proto-instance is a tuple which contains the component it represents, its init argument (which can be anything), and a list of destinations, arranged by out port.

t()

t() :: %Skitter.Workflow{
  instances: %{required(workflow_identifier()) => proto_instance()},
  sources: %{required(workflow_identifier()) => [destination()]}
}

Workflow data structure.

A workflow is defined as a combination of named sources and proto-instances.

token()

token() :: {any(), destination()}

Data record with a destination.

workflow_identifier()

workflow_identifier() :: atom()

Identifier of an instance or source in a workflow.

Functions

get_component!(workflow, key)

get_component!(t(), workflow_identifier()) ::
  Skitter.Component.t() | no_return()

Retrieve the component of a proto-instance based on its identifier.

Examples

iex> get_component!(example_workflow(), :i1)
Identity
iex> get_component!(example_workflow(), :does_not_exist)
** (KeyError) Key `:does_not_exist` not found in workflow

get_init!(workflow, key)

get_init!(t(), workflow_identifier()) :: any() | no_return()

Retrieve the init argument of a proto-instance based on its identifier.

Examples

iex> get_init!(example_workflow(), :i1)
nil

get_links!(workflow, key)

get_links!(t(), workflow_identifier()) :: outgoing_links() | no_return()

Retrieve the links of a proto-instance based on its identifier.

Examples

iex> get_links!(example_workflow(), :i1)
[value: [i3: :value]]

init_proto_instance(workflow, key)

init_proto_instance(t(), workflow_identifier()) ::
  {:ok, Skitter.Component.instance()}
  | Skitter.Component.runtime_error()
  | no_return()

Initialize the proto_instance at key

Examples

iex> init_proto_instance(example_workflow(), :i1)
{:ok, %Skitter.Component.Instance{component: WorkflowTest.Identity, state: []}}

source_data_to_tokens!(workflow, source_data)

source_data_to_tokens!(t(), [{workflow_identifier(), any()}]) :: [token()]

Generate tokens out of the initial data for a workflow.

The initial data is provided as a keyword list, where the keys are source names and the values are values to be sent to the sources. There should be a value for each source; this is currently not verified.

Examples

iex> source_data_to_tokens!(example_workflow(), [s1: 20, s2: 22])
[{20, {:i1, :value}}, {22, {:i2, :value}}]
iex> source_data_to_tokens!(example_workflow(), [does_not_exist: 20])
** (KeyError) Key `:does_not_exist` not found in workflow

spits_to_tokens!(workflow, key, spits)

spits_to_tokens!(t(), workflow_identifier(), [
  {Skitter.Component.port_name(), any()}
]) :: [token()]

Generate tokens from spits like spits_to_tokens/2, but fetch the links.

Identical to spits_to_tokens/2, but fetches the links from a workflow first.

Examples

iex> spits_to_tokens!(example_workflow(), :i1, [value: 20])
[{20, {:i3, :value}}]

spits_to_tokens(links, spits)

spits_to_tokens(outgoing_links(), [{Skitter.Component.port_name(), any()}]) ::
  [token()]

Create a list of tagged tokens based on the outputs of a component instance.

Spits is the data that the component produced while reacting; they are represented as a keyword list where a key is an out port while the value is the data spit to that output port.

Tokens are data records that will be sent to other components in the workflow. Each token is tagged with its destination.

This function receives a list of spits, along with the list of outgoing links of a component instance. Based on these, it creates a list of tagged tokens.

Examples

iex> links = get_links!(example_workflow(), :i1)
[value: [i3: :value]]
iex> spits_to_tokens(links, [value: 20])
[{20, {:i3, :value}}]

workflow(list) (macro)

Create a workflow.

This macro is a shorthand for the Skitter.Workflow.DSL.workflow/1 macro, which enables the creation of skitter workflows. Please refer to the documentation of the Skitter.Workflow.DSL.