Skitter v0.1.0-dev Skitter.Workflow

Tools to interact with skitter workflows.

This module defines an API that can be used to fetch data from skitter workflow structures. Besides this, this module defines the workflow/1 macro which can be used to define a skitter workflow.

Types

destination()

destination() :: {workflow_identifier(), Skitter.Component.port_name()}

Address a token can be sent to.

Specified as the combination of a valid workflow identifier that refers to a component instance and the name of a valid port of that component.

outgoing_links()

outgoing_links() :: [{Skitter.Component.port_name(), [destination()]}]

Outgoing links of a component

Specified as a keyword list. Each key in the list corresponds to an out port of the component, while the values for this key correspond to all the t:destination ports this out port is connected with.

proto_instance()

proto_instance() :: {Skitter.Component.t(), any(), outgoing_links()}

Proto-instance type.

A proto-instance is a tuple which contains the component it represents, its init argument (which can be anything), and a list of destinations, arranged by out port.

t()

t() :: %Skitter.Workflow{
  instances: %{required(workflow_identifier()) => proto_instance()},
  sources: %{required(workflow_identifier()) => [destination()]}
}

Workflow data structure.

A workflow is defined as a combination of named sources and proto-instances.

workflow_identifier()

workflow_identifier() :: atom()

Identifier of an instance or source in a workflow.

Functions

get_component!(workflow, key)

get_component!(t(), workflow_identifier()) ::
  Skitter.Component.t() | no_return()

Retrieve the component of a proto-instance based on its identifier.

Examples

iex> get_component!(example_workflow(), :i1)
Identity
iex> get_component!(example_workflow(), :does_not_exist)
** (KeyError) Key `:does_not_exist` not found in workflow

get_init!(workflow, key)

get_init!(t(), workflow_identifier()) :: any() | no_return()

Retrieve the init argument of a proto-instance based on its identifier.

Examples

iex> get_init!(example_workflow(), :i1)
nil

get_links!(workflow, key)

get_links!(t(), workflow_identifier()) :: outgoing_links() | no_return()

Retrieve the links of a proto-instance based on its identifier.

Examples

iex> get_links!(example_workflow(), :i1)
[value: [i3: :value]]

get_source!(workflow, key)

get_source!(t(), workflow_identifier()) :: [destination()] | no_return()

Obtain the destinations of a source based on its name.

Examples

iex> get_source!(example_workflow(), :s1)
[i1: :value]
iex> get_source!(example_workflow(), :foo)
** (KeyError) Key `:foo` not found in workflow

get_sources(workflow)

get_sources(t()) :: [workflow_identifier()]

List the sources of a component.

Examples

iex> get_sources(example_workflow())
[:s1, :s2]

init_proto_instance!(workflow, key)

init_proto_instance!(t(), workflow_identifier()) ::
  {:ok, Skitter.Component.instance()}
  | Skitter.Component.runtime_error()
  | no_return()

Initialize the proto_instance at key

Examples

iex> init_proto_instance!(example_workflow(), :i1)
{:ok, %Skitter.Component.Instance{component: WorkflowTest.Identity, state: []}}

sources_match?(workflow, kw_list)

Check if a keyword lists matches with the sources of a workflow.

This means that the keys of the keyword list and source names should be identical. Ordering is not taken into account.

Examples

iex> sources_match?(example_workflow(), s1: 3, s2: 4)
true
iex> sources_match?(example_workflow(), s1: 3, s1: 4)
false
iex> sources_match?(example_workflow(), s1: 3)
false

workflow(list) (macro)

Create a workflow.

This macro is a shorthand for the Skitter.Workflow.DSL.workflow/1 macro, which enables the creation of skitter workflows. Please refer to the documentation of the Skitter.Workflow.DSL.