Skitter v0.1.0-dev Skitter.Workflow.DSL

DSL to define skitter workflows.

This module defines the workflow/1 macro, which can be used to create a Skitter workflow. This macro is also exported by the Skitter.Workflow module.

A Skitter workflow is defined by a set of uniquely named proto-instances and sources. Proto-instances define the components that will process the data flowing through a workflow, while sources determine how data that enters the workflow is sent to the proto-instances.

As an example, let's look at a workflow which uses the components Foo and Bar. Foo has in ports a and b, and out ports x and y; Bar only has a single in port: val.

import Skitter.Workflow

workflow do
  source s ~> {f1.a, f1.b}

  f1 = {Foo, {:some_atom, 5}, x ~> f2.a, y ~> f2.b}
  f2 = {Foo, {:some_atom, 4}, y ~> bar.val}

  bar = {Bar, _}
end

This workflow consists of two proto-instances of the same component: Foo, a proto-instance of the component Bar and a single source s which links to Foo.

Sources

Sources connect a workflow to the external world and are defined with the following syntax:

source <name> ~> {<instance identifier>.<port>, <instance identifier>.<port>}

Thus, a source definition consists of two parts: a name and a list of destinations. A destination is specified as the combination of a proto-instance identifier and the name of a valid in port of this proto-instance. When a source only links to a single destination, the curly braces may be omitted: source s ~> foo.bar.

A workflow may have multiple sources.

TODO: Provide additional info when runtime documentation is ready

Proto-instances

A proto-instance is a tuple which contains all the data that is needed to initialize a proper component instance. A proto-instance is defined with the following syntax:

<name> = {
  <Component>,
  <init argument>,
  <out_port> ~> <instance identifier>.<port>,
  <out_port> ~> <instance identifier>.<port>
}

A proto-instance is defined by a Skitter component, its initialization argument and by its links:

The component determines the component that will be instantiated inside the workflow.

The initialization argument will be passed to the Skitter.Component.init/2 function alongside the component to initialize the component instance. It can be used to configure the component before it is initialized. _ can be used when no arguments need to be passed.

The links determine to which proto-instances this proto-instance is connected. A link is specified as a combination of an out port on the proto-instance that is being defined, a proto-instance identifier and a valid in port on this proto-instance.

Functions

workflow(list) (macro)

Create a skitter workflow.

This macro serves as the entry point of the Skitter.Workflow.DSL DSL. Please refer to the module documentation for additional details.